+++ backend/src/@types/requestError.d.ts
@@ -1,5 +1,4 @@
 export interface RequestError extends Error {
   status: number;
-  additionalInformation?: string;
   expected: boolean;
 }