+++ backend/src/prisma/migrations/20260115135016_add_upload_requests/migration.sql
@@ -1,16 +1,3 @@
-/*
-  Warnings:
-
-  - Made the column `isPinned` on table `event` required. This step will fail if there are existing NULL values in that column.
-  - Made the column `isPinned` on table `homework` required. This step will fail if there are existing NULL values in that column.
-
-*/
--- AlterTable
-ALTER TABLE "event" ALTER COLUMN "isPinned" SET NOT NULL;
-
--- AlterTable
-ALTER TABLE "homework" ALTER COLUMN "isPinned" SET NOT NULL;
-
 -- CreateTable
 CREATE TABLE "uploadRequest" (
     "uploadRequestId" SERIAL NOT NULL,