+++ backend/src/utils/validate.functions.ts
@@ -3,7 +3,6 @@ import { CACHE_KEY_PREFIXES, cacheExpiration, generateCacheKey, redisClient } fr
 import prisma from "../config/prisma";
 import logger from "../config/logger";
 import { Session, SessionData } from "express-session";
-import { FileTypes } from "../config/upload";
 
 async function updateCacheData<T>(data: T[], key: string): Promise<void> {
   try {
@@ -35,19 +34,6 @@ export function checkUsername(username: string): boolean {
 
 function BigIntreplacer(key: string, value: unknown): unknown {
   return typeof value === "bigint" ? value.toString() : value;
-}
-
-async function isValidUploadInput(uploadName: string, uploadDescription: string | null, uploadType: string): Promise<void> {
-  const namePassing = uploadName !== "" && uploadDescription !== "";
-  if (!(namePassing && Object.values(FileTypes).includes(uploadType as FileTypes))) {
-    const err: RequestError = {
-      name: "Bad Request",
-      status: 400,
-      message: "Please provide a valid name, teamId (int) and valid file type (INFO_SHEET,LESSON_NOTE,WORKSHEET,IMAGE,FILE,TEXT)",
-      expected: true
-    };
-    throw err;
-  }
 }
 
 async function isValidEventTypeId(eventTypeId: number, session: Session & Partial<SessionData>): Promise<void> {
@@ -177,7 +163,6 @@ function lessonDateEventAtLeastOneNull(endDate: number | null, lesson: string | 
 }
 
 export {
-  isValidUploadInput,
   isValidColor,
   isValidSubjectId,
   isValidTeamId,